backend/requirements.txt
Django>=5.0
djangorestframework
psycopg2-binary
django-cors-headers
gunicorn
uvicorn
whitenoise
black
flake8
